services/geoip_enricher/requirements.txt
kafka-python==2.0.2
geoip2==4.7.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1